+++ R/geom_segment_c.R
@@ -54,7 +54,6 @@ geom_segment_c <- function(mapping = NULL, data = NULL,
 Stat <- getFromNamespace("Stat", "ggplot2")
 
 ##' @importFrom ggplot2 ggplot_add
-##' @importFrom ggplot2 aes_string
 ##' @method ggplot_add segmentC
 ##' @export
 ggplot_add.segmentC <- function(object, plot, object_name, ...) {
@@ -69,7 +68,7 @@ ggplot_add.segmentC <- function(object, plot, object_name, ...) {
     mapping <- object$mapping
     # mapping["colour"] <- list(v)
 
-    default_aes <- aes_string(colour=v)
+    default_aes <- aes(colour = !!rlang::sym(v))
     if (is.null(mapping)) {
         mapping <- default_aes
     } else {
@@ -127,14 +126,14 @@ setup_data_continuous_color_df <- function(df, nsplit = 100, extend = 0.002, poo
                                            extend = extend)
 
         res <- lapply(df[i,, drop = FALSE], rep, each = nrow(df2)) |>
-            do.call('cbind', args = list()) |> as.data.frame()
+            do.call(what = 'cbind') |> as.data.frame()
         res$x <- df2$x
         res$xend <- df2$xend
         res$y <- df2$y
         res$yend <- df2$yend
-        res$colour <- df2$col
+        res$colour <- df2$colour
         return(res)
-    }) |> do.call('rbind', args = list())
+    }) |> do.call(what = 'rbind')
 }
 
 
@@ -146,23 +145,31 @@ setup_data_continuous_color <- function(x, xend, y, yend, col, col2,
     if (is.null(xrange))
         xrange <- c(x, xend)
 
-    ## xstep <- diff(xrange)/nsplit
-    ## xn <- floor((xend - x)/xstep)
-    xn <- floor((xend - x) * nsplit /diff(xrange))
-    ## slope <- (yend - y)/(xend - x)
     ydiff <- yend - y
     xdiff <- xend - x
 
-    if (xn > 0) {
-        ## x <- x + 0:xn * xstep
-        x <- x + 0:xn * diff(xrange) / nsplit 
-        tmp <- x[-1] * (1 + extend)
-        tmp[tmp > xend] <- xend
-        xend <- c(tmp, xend)
-        ## y <- y + 0:xn * xstep * slope
-        y <- y + 0:xn * diff(xrange) * ydiff / (nsplit * xdiff)
-        ## yend <- y + (xend - x) * slope
-        yend <- y + (xend - x) * ydiff / xdiff 
+    if (xdiff == 0 && ydiff == 0) {
+        n <- 1
+    } else {
+        span <- if (xdiff == 0) abs(ydiff) else abs(diff(xrange))
+        if (span == 0) {
+            span <- if (xdiff == 0) abs(ydiff) else abs(xdiff)
+        }
+        n <- max(1, floor(max(abs(xdiff), abs(ydiff)) * nsplit / span))
+    }
+
+    if (n > 1) {
+        t0 <- seq(0, 1, length.out = n)
+        t1 <- c(t0[-1] * (1 + extend), 1)
+        t1[t1 > 1] <- 1
+        t1[t1 < 0] <- 0
+
+        x0 <- x
+        y0 <- y
+        x <- x0 + t0 * xdiff
+        xend <- x0 + t1 * xdiff
+        y <- y0 + t0 * ydiff
+        yend <- y0 + t1 * ydiff
     }
 
     n <- length(x)